feat(util) - move util methods to separate file. Case: use loadFileConfigs in default.js to load parent config from other folder BEFORE new Config() occurs
