+++ .gitignore
@@ -1,3 +1,4 @@
+.idea/
 .project
 node_modules
 npm-debug.log
