+++ package.json
@@ -45,6 +45,6 @@
     "node": ">= 4.0.0"
   },
   "scripts": {
-    "test": "./node_modules/vows/bin/vows test/*.js --spec"
+    "test": "vows test/*.js --spec"
   }
 }
